Basic framework set up
I think this encompasses all or most of the Model part of MVC.
The Dao class contains a Main method that I've been using for testing
purposes.

Init() loads the SQL file to build and populate tables. It assumes
there is a MySQL database called 'infprg03' and a user with name and
password of 'root'.

All other  commented statements should work, using only System.out for
now. Some TODO's are marked, only a few of which are actually critical
as far as I can tell.